third_party/blink/web_tests/svg/custom/polyline-setattribute-points-null-expected.txt

CONSOLE ERROR: Error: <polyline> attribute points: Expected number, "undefined".
